+++ internal/scanner/hashutil.go
@@ -1,0 +1,11 @@
+package scanner
+
+import (
+	"crypto/sha256"
+	"encoding/hex"
+)
+
+func hashBytesSHA256(data []byte) string {
+	sum := sha256.Sum256(data)
+	return hex.EncodeToString(sum[:])
+}